% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{check_data}
\alias{check_data}
\alias{check_header}
\title{Deprecated functions kept for compatibility with the lidR package versions <= 2.0.0}
\usage{
check_data(data)

check_header(header)
}
\arguments{
\item{data}{a data.frame or a data.table containing a point cloud}

\item{header}{a list containing the header of a las file}

\item{...}{internal use only}
}
\description{
Deprecated functions kept for compatibility with the lidR package versions <= 2.0.0
}
