% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLAS.R
\name{lasfile}
\alias{lasfile}
\alias{lasfile.character}
\title{Read LiDAR LAS file}
\usage{
lasfile(x, ...)

\method{lasfile}{character}(x, skip = 0, nrows = NULL,
  returnHeaderOnly = FALSE, ...)
}
\arguments{
\item{x}{LAS file to read}

\item{...}{reserved args, unused}

\item{skip}{number of records to skip}

\item{nrows}{number of records to read}

\item{returnHeaderOnly}{logical, return header only?}
}
\description{
Read LiDAR LAS file
}

