% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_tools.r
\name{make_header}
\alias{make_header}
\title{Create a header from data}
\usage{
make_header(data, extra_bytes = NULL)
}
\arguments{
\item{data}{data.frame or data.table}

\item{extra_bytes}{Labelled list. Specify which column must be saved as extra bytes attribute and
provide a description to this data. The}
}
\description{
Create a header for a las file from data. The header is not necessarily entierely correct regarding
specification but is expected to be enought to be passed in \link{write.las} since many redoundant
information are computed at writing time anyway.
}
\examples{
lasdata = data.frame(X = c(339002.889, 339002.983, 339002.918),
                     Y = c(5248000.515, 5248000.478, 5248000.318),
                     Z = c(975.589, 974.778, 974.471),
                     gpstime = c(269347.281418006, 269347.281428006, 269347.281438006),
                     Intensity = c(82L, 54L, 27L),
                     ReturnNumber = c(1L, 1L, 2L),
                     NumberOfReturns = c(1L, 1L, 2L),
                     ScanDirectionFlag = c(1L, 1L, 1L),
                     EdgeOfFlightline = c(1L, 0L, 0L),
                     Classification = c(1L, 1L, 1L),
                     ScanAngle = c(-21L, -21L, -21L),
                     UserData = c(32L, 32L, 32L),
                     PointSourceID = c(17L, 17L, 17L),
                     treeID = c(1L, 1L, 1L))

extra = list(treeID = "Unique ID for segmented trees")

lasheader = make_header(lasdata, extra)
}
