% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readLAS.R
\name{readLAS}
\alias{readLAS}
\title{rlas}
\usage{
readLAS(lasfile, skip = 0, nrows = NULL, returnHeaderOnly = FALSE)
}
\arguments{
\item{lasfile}{LAS file to read}

\item{skip}{number of records to skip}

\item{nrows}{number of records to read}

\item{returnHeaderOnly}{logical, return header only?}
}
\description{
rlas
}

