% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{readlasheader}
\alias{readlasheader}
\title{(deprecated) Read header from a .las or .laz file}
\usage{
readlasheader(file)
}
\arguments{
\item{file}{filepath character string to the .las or .laz file}
}
\value{
A \code{list}
}
\description{
(deprecated) Use \link{read.lasheader}
}
\examples{
lazfile   <- system.file("extdata", "example.laz", package="rlas")
lasheader <- readlasheader(lazfile)
}
