% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.r
\name{writelas}
\alias{writelas}
\title{(deprecated) Write a .las or .laz file}
\usage{
writelas(file, header, X, Y, Z, gpstime, Intensity, ReturnNumber,
  NumberOfReturns, ScanDirectionFlag, EdgeOfFlightline, Classification,
  ScanAngle, UserData, PointSourceID, R, G, B)
}
\arguments{
\item{file}{character. filename of .las or .laz file}

\item{header}{list of character. The data for the file header properly labelled
(see \link[rlas:readlasheader]{readlasheader})}

\item{X}{numeric array X data}

\item{Y}{numeric array Y data}

\item{Z}{numeric array Z data}

\item{gpstime}{numeric array gpstime data}

\item{Intensity}{integer array intensity data}

\item{ReturnNumber}{integer array return number data}

\item{NumberOfReturns}{integer array number of returns data}

\item{ScanDirectionFlag}{integer array scan direction flag data}

\item{EdgeOfFlightline}{integer array edge of flightline data}

\item{Classification}{integer array classification data}

\item{ScanAngle}{integer array scan angle data}

\item{UserData}{integer array user data data}

\item{PointSourceID}{integer array point source id data}

\item{R}{integer array red data}

\item{G}{integer array green data}

\item{B}{integer array blue data}
}
\value{
void
}
\description{
(deprecated) Use \link{write.las}
}
