% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.r
\name{check_data}
\alias{check_data}
\alias{check_header}
\alias{check_data_vs_header}
\title{Check if data and headers are correct}
\usage{
check_data(data)

check_header(header)

check_data_vs_header(header, data, ...)
}
\arguments{
\item{data}{a data.frame or a data.table containing a point cloud}

\item{header}{a list containing the header of a las file}

\item{...}{internal use only}
}
\description{
las files are normalized files. The \code{check_*} functions test if the data is valid,
if the header is valid or if the header is in accordance with the data.
}
