% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{compression}
\alias{compression}
\alias{is_compressed}
\alias{true_size}
\title{ALTREP utilities}
\usage{
is_compressed(x)

true_size(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\code{is_compressed} return a boolean. \code{true_size} return the real size of the object in bytes
}
\description{
Test if an a vector is compressed using the ALTREP framework
}
\examples{
\dontrun{
lazfile <- system.file("extdata", "example.las", package = "rlas")
las <- read.las(lazfile)
is_compressed(las)

# The difference is more substantial on bigger point clouds (~30\%)
object.size(las)
true_size(las)
}
}
