% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_tools.r
\name{crs_tools}
\alias{crs_tools}
\alias{header_get_epsg}
\alias{header_set_epsg}
\alias{header_get_wktcs}
\alias{header_set_wktcs}
\alias{find_epsg_position}
\title{Coordinate Reference System Tools}
\usage{
header_get_epsg(header)

header_set_epsg(header, epsg)

header_get_wktcs(header)

header_set_wktcs(header, WKT)

find_epsg_position(header)
}
\arguments{
\item{header}{list}

\item{epsg}{integer. An EPSG code}

\item{WKT}{string. A string of an WKT OGC CS}
}
\description{
Functions that update a header to describe coordinates reference system according to the
\href{https://community.asprs.org/leadership-restricted/leadership-content/public-documents/standards}{LAS specifications}
}
\seealso{
Other header_tools:
\code{\link{extra_bytes_attribute_tools}},
\code{\link[=fwf_interpreter]{fwf_interpreter()}},
\code{\link{public_header_block_tools}}
}
\concept{header_tools}
