% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwf_interpreter.r
\name{fwf_interpreter}
\alias{fwf_interpreter}
\title{Full Waveform interpreter}
\usage{
fwf_interpreter(header, data)
}
\arguments{
\item{header}{list. A header}

\item{data}{data.frame or data.table}
}
\value{
A list containing a \code{data.frame} per points
}
\description{
Full waveform data read from las files might be cryptic even with a good understanding of the LAS
specifications. This function interpret the Full waveform data in a set of XYZ coordinates and and
amplitude normalized by the maximum value possible (i.e 2 power the number of bits used to store
the full waveform)
}
\seealso{
Other header_tools: 
\code{\link{crs_tools}},
\code{\link{extra_bytes_attribute_tools}},
\code{\link{public_header_block_tools}}
}
\concept{header_tools}
