% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLAS.r
\name{readlasdata}
\alias{readlasdata}
\title{Read data from a .las or .laz file}
\usage{
readlasdata(file, Intensity = TRUE, ReturnNumber = TRUE,
  NumberOfReturns = TRUE, ScanDirectionFlag = TRUE,
  EdgeOfFlightline = TRUE, Classification = TRUE, ScanAngle = TRUE,
  UserData = TRUE, PointSourceID = TRUE, RGB = TRUE, filter = "")
}
\arguments{
\item{file}{characters path to the .las or .laz file}

\item{Intensity}{logical. do you want to load the Intensity field? default: TRUE}

\item{ReturnNumber}{logical. do you want to load the ReturnNumber field? default: TRUE}

\item{NumberOfReturns}{logical. do you want to load the NumberOfReturns field? default: TRUE}

\item{ScanDirectionFlag}{logical. do you want to load the ScanDirectionFlag field? default: TRUE}

\item{EdgeOfFlightline}{logical. do you want to load the EdgeOfFlightline field? default: TRUE}

\item{Classification}{logical. do you want to load the Classification field? default: TRUE}

\item{ScanAngle}{logical. do you want to load the ScanAngle field? default: TRUE}

\item{UserData}{logical. do you want to load the UserData field? default: TRUE}

\item{PointSourceID}{logical. do you want to load the PointSourceID field? default: TRUE}

\item{RGB}{logical. do you want to load R,G and B fields? default: TRUE}

\item{filter}{character. filter data while reading the file (streaming filter) without
allocating any useless memory. (see Details).}
}
\value{
A \code{data.table}
}
\description{
Read data from .las or .laz files in format 1 to 4 according to LAS specifications and returns
a \code{data.table} labeled according to LAS specifications. See the ASPRS documentation for the
\href{http://www.asprs.org/a/society/committees/standards/LAS_1_4_r13.pdf}{LAS file format}.
The optional logical parameters enable the user to save memory by choosing to load only the
fields they need. Indeed, data is loaded into the computer's memory (RAM) suboptimally because
R does not accommodate many different data types. Moreover the function provides a streaming filter
to load only the points of interest into the memory never allocating any usless memory.
}
\details{
Because \code{rlas} rely on the well know \code{LASlib} library written by Martin Isenburg
to read the binary files, the package inherit also of filter commands avaible in
\href{https://rapidlasso.com/lastools/}{LAStools}. To use these filters the user can pass the
common commands from \code{LAStools} into the parameter \code{'filter'}. Type \code{rlas:::lasfilterusage()} to
display the documentation of \code{LASlib} and the avaible filters.\cr\cr
The filter works in two passes. First it streams the file without loading anything and count
the number of points of interest. Then it allocates the necessary amount of memory and read the file
a second time and stores the point of interests into the computer's memory (RAM).
}
\examples{
lazfile <- system.file("extdata", "example.laz", package="rlas")

lasdata <- readlasdata(lazfile)
lasdata <- readlasdata(lazfile, filter = "-keep_first")
lasdata <- readlasdata(lazfile, filter = "-drop_intensity_below 80")
}
\seealso{
Other rlas: \code{\link{readlasheader}},
  \code{\link{writelas}}
}

