% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLAX.r
\name{writelax}
\alias{writelax}
\title{Write a .lax file from a .las or .laz file}
\usage{
writelax(file)
}
\arguments{
\item{file}{character. filename of .las or .laz file}
}
\description{
Write a lax file from a las or laz file. A lax file is a tiny file which can come with a
las or laz and which spatially index the data to make faster spatial queries. It has
been invented by Martin Isenburg in LASlib. rlas support lax file and enable to write
a lax file with default settings. For more options, use lasindex from binaries provided
by LASlib (for more informations see references)
}
\references{
https://rapidlasso.com/\cr
https://rapidlasso.com/2012/12/03/lasindex-spatial-indexing-of-lidar-data/\cr
https://github.com/LAStools/LAStools
}
